created src directory and moved client and server folders into it, keeping server.js entry at root directory
