+++ package.json
@@ -7,9 +7,9 @@
     "start": "if-env NODE_ENV=production && yarn start:prod || yarn start:dev",
     "start:prod": "node server.js",
     "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"yarn client\"",
-    "client": "cd client && yarn start",
-    "install": "cd client && yarn install",
-    "build": "cd client && yarn build",
+    "client": "cd src/client && yarn start",
+    "install": "cd src/client && yarn install",
+    "build": "cd src/client && yarn build",
     "heroku-postbuild": "yarn build"
   },
   "author": "Vincent Shury",
