+++ server.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const mongoose = require('mongoose')
-const routes = require('./routes')
+const routes = require('./src/routes')
 const app = express()
 const PORT = process.env.PORT || 3001
 const MONGODB_URI = process.env.MONGODB_URI || 'mongodb://localhost/googlebooks'
@@ -11,7 +11,7 @@ app.use(express.json())
 mongoose.connect(MONGODB_URI, mongooseConfig, error => console.log(error || '--> Connected to Database'))
 
 if (process.env.NODE_ENV === 'production') {
-  app.use(express.static('client/build'))
+  app.use(express.static('src/client/build'))
 }
 
 app.use(routes)
