+++ src/client/public/manifest.json
