+++ src/client/src/components/DeleteBtn/index.js
